Harvest property for harvest amount changed from double to decimal for more precision

+++ BeeSoft.Data/Models/Harvest.cs
@@ -12,7 +12,7 @@ public class Harvest
 
     public DateTime HarvestDate { get; init; }
 
-    public double HarvestedAmount { get; init; }
+    public decimal HarvestedAmount { get; init; }
 
     [Required]
     [MaxLength(HarvestedProductMaxLength)]
