+++ BeeSoft.Web/Models/Harvests/UpdateHarvestFormModel.cs
@@ -13,7 +13,7 @@ public sealed record UpdateHarvestFormModel
 
     public DateTime HarvestDate { get; init; }
 
-    public double HarvestedAmount { get; init; }
+    public decimal HarvestedAmount { get; init; }
 
     [Required]
     [MaxLength(HarvestedProductMaxLength)]
